Make configurable products always have stock if at least one of the children has stock (configurables don't have inventory source items). Also add trait for removing duplicate code
